+++ inst/python/huggingface_Interface3.py
@@ -50,7 +50,7 @@ def set_logging_level(logging_level):
     if logging_level in ['warn', 'warning']:
         logging.set_verbosity_warning()
     elif logging_level == "critical":
-        logging.set_verbosity_critical()
+        logging.set_verbosity(50)
     elif logging_level == "error":
         logging.set_verbosity_error()
     elif logging_level == "info":
@@ -74,7 +74,7 @@ def get_device(device):
     Parameters
     ----------
     device : str
-        name of device: 'cpu', 'gpu', 'cuda', or of the form 'gpu:k' or 'cuda:k' 
+        name of device: 'cpu', 'gpu', 'cuda', 'mps', or of the form 'gpu:k', 'cuda:k', or 'mps:0'
         where k is a specific device number
 
     Returns
@@ -85,8 +85,8 @@ def get_device(device):
         device number, -1 for CPU
     """
     device = device.lower()
-    if not device.startswith('cpu') and not device.startswith('gpu') and not device.startswith('cuda'):
-        print("device must be 'cpu', 'gpu', 'cuda', or of the form 'gpu:k' or 'cuda:k'")
+    if not device.startswith('cpu') and not device.startswith('gpu') and not device.startswith('cuda') and not device.startswith('mps'):
+        print("device must be 'cpu', 'gpu', 'cuda', 'mps', or of the form 'gpu:k', 'cuda:k', or 'mps:0'")
         print("\twhere k is an integer value for the device")
         print("Trying CPUs")
         device = 'cpu'
@@ -94,21 +94,38 @@ def get_device(device):
     device_num = -1
     if device != 'cpu':
         attached = False
+        
+        if hasattr(torch.backends, "mps"):
+            mps_available = torch.backends.mps.is_available()
+        else:
+            mps_available = False
+        print(f"MPS for Mac available: {mps_available}")
         if torch.cuda.is_available():
             if device == 'gpu' or device == 'cuda': 
                 # assign to first gpu device number
                 device = 'cuda'
                 device_num = list(range(torch.cuda.device_count()))[0]
                 attached = True
-            else: # assign to specific gpu device number
+        elif "mps" in device:
+            if not torch.backends.mps.is_available():
+                if not torch.backends.mps.is_built():
+                    print("MPS not available because the current PyTorch install was not built with MPS enabled.")
+                else:
+                    print("MPS not available because the current MacOS version is not 12.3+ and/or you do not have an MPS-enabled device on this machine.")
+            else:
+                device_num = 0 # list(range(torch.cuda.device_count()))[0]
+                device = 'mps:' + str(device_num)
+                attached = True
+                print("Using Metal Performance Shaders (MPS) backend for GPU training acceleration!")
+        else: # assign to specific gpu device number
                 try:
                     device_num = int(device.split(":")[-1])
                     device = 'cuda:' + str(device_num)
                     attached = True
                 except:
-                    pass
+                    attached = False
         if not attached:
-            print("Unable to use CUDA (GPU), using CPU")
+            print("Unable to use MPS (Mac M1+), CUDA (GPU), using CPU")
             device = "cpu"
             device_num = -1
 
@@ -471,7 +488,8 @@ def hgTransformerGetTranslation(text_strings,
                             target_lang = '',
                             return_tensors = False,
                             return_text = True,
-                            clean_up_tokenization_spaces = False):
+                            clean_up_tokenization_spaces = False, 
+                            max_length = ''):
     task = 'translation'
     if source_lang and target_lang:
         task = "translation_{s}_to_{t}".format(s=source_lang, t=target_lang)
@@ -487,7 +505,8 @@ def hgTransformerGetTranslation(text_strings,
                             tgt_lang = target_lang,
                             return_tensors = return_tensors,
                             return_text = return_text,
-                            clean_up_tokenization_spaces = clean_up_tokenization_spaces)
+                            clean_up_tokenization_spaces = clean_up_tokenization_spaces, 
+                            max_length = max_length)
     return translations
 
 def hgTransformerGetEmbedding(text_strings,
@@ -574,7 +593,7 @@ def hgTransformerGetEmbedding(text_strings,
             if return_tokens:
                 tokens = []
                 for ids in input_ids:
-                    tokens.extend([token for token in tokenizer.convert_ids_to_tokens(ids) if token != '[PAD]'])
+                    tokens.extend([token for token in tokenizer.convert_ids_to_tokens(ids) if token != '[PAD]' and token != '<pad>'])
                 all_toks.append(tokens)
 
             with torch.no_grad():
@@ -685,7 +704,7 @@ def hgTokenizerGetTokens(text_strings,
 
             tokens = []
             for ids in input_ids:
-                tokens.extend([token for token in tokenizer.convert_ids_to_tokens(ids) if token != '[PAD]'])
+                tokens.extend([token for token in tokenizer.convert_ids_to_tokens(ids) if token != '[PAD]' and token != '<pad>'])
             all_toks.append(tokens)
 
         else: